Refactor ShSimulatorService
